+++ transport.go
@@ -13,6 +13,9 @@
 //
 //	rec := recorder.NewMemoryRecorder()
 //	config := recorder.DefaultConfig()
+//	if err := config.Validate(); err != nil {
+//		log.Fatal(err)
+//	}
 //	client := &http.Client{
 //		Transport: recorder.NewTransport(http.DefaultTransport, rec, config),
 //	}